Integrate `keycdn` plugin into `site` plugin
